refactor: Rename endpoint routers
To better clarify the use of the routers for existing BE endpoints and
to avoid naming collisions they should be renamed like `TcsServiceRouter.

TISNEW-4785
