+++ build.gradle
@@ -32,8 +32,9 @@ dependencies {
   }
 
   // Apache Camel starters
-  implementation "org.apache.camel.springboot:camel-spring-boot-starter:3.4.0"
-  implementation "org.apache.camel.springboot:camel-http-starter:3.4.0"
+  ext.camelVersion = "3.4.0"
+  implementation "org.apache.camel.springboot:camel-spring-boot-starter:$camelVersion"
+  implementation "org.apache.camel.springboot:camel-http-starter:$camelVersion"
 
   // Lombok
   compileOnly "org.projectlombok:lombok"
