+++ src/main/java/uk/nhs/hee/tis/revalidation/integration/router/service/CoreServiceRouter.java
@@ -19,14 +19,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package uk.nhs.hee.tis.revalidation.integration.router;
+package uk.nhs.hee.tis.revalidation.integration.router.service;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
-public class CoreRouter extends RouteBuilder {
+public class CoreServiceRouter extends RouteBuilder {
 
   private static final String API_DOCTORS = "/api/doctors";
