+++ src/main/java/uk/nhs/hee/tis/revalidation/integration/router/service/TcsServiceRouter.java
@@ -19,14 +19,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package uk.nhs.hee.tis.revalidation.integration.router;
+package uk.nhs.hee.tis.revalidation.integration.router.service;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
-public class TcsRouter extends RouteBuilder {
+public class TcsServiceRouter extends RouteBuilder {
 
   private static final String API_TRAINEE =
       "/api/revalidation/trainee/${header.gmcId}?bridgeEndpoint=true";
